+++ packages/spotlight/src/spotlight.js
@@ -717,8 +717,8 @@ const Spotlight = (function () {
 		 * If Spotlight is in pointer mode, focus is not changed but `elem` will be set as the last
 		 * focused element of its spotlight containers.
 		 *
-		 * @param {String|Node|undefined} elem The spotlight ID or selector for either a spottable
-		 *  component or a spootliht container, or spottable node. If not supplied, the default
+		 * @param {String|Node} [elem] The spotlight ID or selector for either a spottable
+		 *  component or a spotlight container, or spottable node. If not supplied, the default
 		 *  container will be focused.
 		 * @returns {Boolean} `true` if focus successful, `false` if not.
 		 * @public