+++ packages/core/handle/handle.js
@@ -82,6 +82,20 @@ import curry from 'ramda/src/curry';
 
 import {is} from '../keymap';
 
+/**
+ * @callback EventHandler
+ * @memberof core/handle
+ * @param {any} event
+ */
+
+/**
+ * @callback HandlerFunction
+ * @memberof core/handle
+ * @param {Object<string, any>} event
+ * @param {any} props
+ * @param {Object<string, any>} context
+ */
+
 // Accepts an array of handlers, sanitizes them, and returns a handler function
 // compose(allPass, map(makeSafeHandler));
 const makeHandler = (handlers) => {
@@ -149,9 +163,9 @@ const decorateHandleFunction = (fn) => {
  * of the returned value.
  *
  * @method   handle
- * @param    {...Function}  handlers List of handlers to process the event.
+ * @param    {...HandlerFunction}  handlers List of handlers to process the event.
  *
- * @returns  {Function}	    A function that accepts an event which is dispatched to each of the
+ * @returns  {EventHandler}	A function that accepts an event which is dispatched to each of the
  *                          provided handlers.
  * @memberof core/handle
  * @public
@@ -222,7 +236,7 @@ const handle = function (...handlers) {
  * @method   oneOf
  * @param    {...Function[]}  handlers List of conditions and handlers to process the event
  *
- * @returns  {Function}	    A function that accepts an event which is dispatched to each of the
+ * @returns  {HandlerFunction} A function that accepts an event which is dispatched to each of the
  *                          conditions and, if it passes, onto the provided handler.
  * @memberof core/handle
  * @public
@@ -237,7 +251,7 @@ const oneOf = handle.oneOf = function (...handlers) {
  *
  * Example:
  * ```
- * // Used to coerce an existing function into a handler change
+ * // Used to coerce an existing function into a handler
  * const coercedHandler = handle(
  *   returnsTrue(doesSomething),
  *   willAlwaysBeCalled
@@ -253,12 +267,12 @@ const oneOf = handle.oneOf = function (...handlers) {
  * @method   returnsTrue
  * @param    {Function}  [handler]  Handler function called before returning `true`.
  *
- * @returns  {Function}	   A function that returns true
+ * @returns  {HandlerFunction}	   A function that returns `true`
  * @memberof core/handle
  * @public
  */
 const returnsTrue = handle.returnsTrue = function (handler) {
-	if (handler) {
+	if (handler && typeof handler === 'function') {
 		return named(function (...args) {
 			handler.apply(this, args);
 
@@ -278,7 +292,7 @@ const returnsTrue = handle.returnsTrue = function (handler) {
  *
  * const callsCustomMethod = handle(
  *	callOnEvent('customMethod'),
- *	(ev) => console.log('ev.customMethod() was called')
+ *	(ev) => console.log('ev.customMethod() was called', ev)
  * );
  * ```
  *
@@ -286,7 +300,8 @@ const returnsTrue = handle.returnsTrue = function (handler) {
  * @param    {String}     methodName  Name of the method to call on the event
  * @param    {Object}     ev          Event payload
  *
- * @returns  {Boolean}                Always returns `true`
+ * @returns  {true}                   Always returns `true`
+ * @curried
  * @memberof core/handle
  * @private
  */
@@ -310,7 +325,7 @@ const callOnEvent = handle.callOnEvent = curry((methodName, ev) => {
  *
  * const logWhenXEqualsZero = handle(
  *   forEventProp('x', 0),
- *   (ev) => console.log('ev.x was equal to zero')
+ *   (ev) => console.log('ev.x was equal to zero', ev)
  * );
  * ```
  *
@@ -320,6 +335,7 @@ const callOnEvent = handle.callOnEvent = curry((methodName, ev) => {
  * @param    {Object}      ev     Event payload
  *
  * @returns  {Boolean}            Returns `true` if `prop` on `event` strictly equals `value`
+ * @curried
  * @memberof core/handle
  * @public
  */
@@ -338,7 +354,7 @@ const forEventProp = handle.forEventProp = curry((prop, value, ev) => {
  *
  * const forwardAndLog = handle(
  *   forward('onClick'),
- *   (ev) => console.log('event forwarded to onClick from props')
+ *   (ev) => console.log('event forwarded to onClick from props', ev)
  * );
  * ```
  *
@@ -347,7 +363,8 @@ const forEventProp = handle.forEventProp = curry((prop, value, ev) => {
  * @param    {Object}    ev     Event payload
  * @param    {Object}    props  Props object
  *
- * @returns  {Boolean}          Always returns `true`
+ * @returns  {true}             Always returns `true`
+ * @curried
  * @memberof core/handle
  * @public
  */
@@ -369,14 +386,14 @@ const forward = handle.forward = curry(named((name, ev, props) => {
  *
  * const preventAndLog = handle(
  *   preventDefault,
- *   (ev) => console.log('preventDefault called')
+ *   (ev) => console.log('preventDefault called', ev)
  * );
  * ```
  *
  * @method   preventDefault
  * @param    {Object}        ev  Event payload
  *
- * @returns  {Boolean}           Always returns `true`
+ * @returns  {true}              Always returns `true`
  * @memberof core/handle
  * @public
  */
@@ -393,7 +410,7 @@ const preventDefault = handle.preventDefault = callOnEvent('preventDefault');
  *
  * const forwardPreventDefault = handle(
  *   forwardWithPrevent('onClick'),
- *   (ev) => console.log('default action')
+ *   (ev) => console.log('default action', ev)
  * );
  * ```
  *
@@ -403,6 +420,7 @@ const preventDefault = handle.preventDefault = callOnEvent('preventDefault');
  * @param    {Object}    props  Props object
  *
  * @returns  {Boolean}          Returns `true` if default action is prevented
+ * @curried
  * @memberof core/handle
  * @private
  */
@@ -428,14 +446,15 @@ const forwardWithPrevent = handle.forwardWithPrevent = curry(named((name, ev, pr
  *
  * const stopAndLog = handle(
  *   stop,
- *   (ev) => console.log('stopPropagation called')
+ *   (ev) => console.log('stopPropagation called', ev)
  * );
  * ```
  *
  * @method   stop
  * @param    {Object}   ev  Event payload
  *
- * @returns  {Boolean}      Always returns `true`
+ * @returns  {true}         Always returns `true`
+ * @curried
  * @memberof core/handle
  * @public
  */
@@ -450,14 +469,15 @@ const stop = handle.stop = named(callOnEvent('stopPropagation'), 'stop');
  *
  * const stopImmediateAndLog = handle(
  *   stopImmediate,
- *   (ev) => console.log('stopImmediatePropagation called')
+ *   (ev) => console.log('stopImmediatePropagation called', ev)
  * );
  * ```
  *
  * @method   stopImmediate
  * @param    {Object}       ev  Event payload
  *
- * @returns  {Boolean}          Always returns `true`
+ * @returns  {true}             Always returns `true`
+ * @curried
  * @memberof core/handle
  * @public
  */
@@ -472,7 +492,7 @@ const stopImmediate = handle.stopImmediate = callOnEvent('stopImmediatePropagati
  *
  * const logForEscapeKey = handle(
  *   forKeyCode(27),
- *   (ev) => console.log('Escape key pressed down')
+ *   (ev) => console.log('Escape key pressed down', ev)
  * );
  * ```
  *
@@ -481,6 +501,7 @@ const stopImmediate = handle.stopImmediate = callOnEvent('stopImmediatePropagati
  * @param    {Object}    ev     Event payload
  *
  * @returns  {Boolean}          Returns `true` if `event.keyCode` strictly equals `value`
+ * @curried
  * @memberof core/handle
  * @public
  */
@@ -496,7 +517,7 @@ const forKeyCode = handle.forKeyCode = forEventProp('keyCode');
  *
  * const logForEnterKey = handle(
  *   forKey('enter'),
- *   (ev) => console.log('Enter key pressed down')
+ *   (ev) => console.log('Enter key pressed down', ev)
  * );
  * ```
  *
@@ -506,6 +527,7 @@ const forKeyCode = handle.forKeyCode = forEventProp('keyCode');
  * @param    {Object}    ev     Event payload
  *
  * @returns  {Boolean}          Returns `true` if `event.keyCode` is mapped to `name`
+ * @curried
  * @memberof core/handle
  * @public
  */
@@ -522,7 +544,7 @@ const forKey = handle.forKey = curry((name, ev) => {
  *
  * const logWhenChecked = handle(
  *   forProp('checked', true),
- *   (ev) => console.log('checked prop is true')
+ *   (ev) => console.log('checked prop is true', ev)
  * );
  * ```
  *
@@ -532,7 +554,8 @@ const forKey = handle.forKey = curry((name, ev) => {
  * @param    {Object}    ev     Event payload
  * @param    {Object}    props  Props object
  *
- * @returns  {Boolean}          `true` if the value of `props[prop]` strictly equals `value`
+ * @returns  {Boolean}          Returns `true` if the value of `props[prop]` strictly equals `value`
+ * @curried
  * @memberof core/handle
  * @public
  */
@@ -559,7 +582,8 @@ const forProp = handle.forProp = curry((prop, value, ev, props) => {
  * @param    {Object}     ev       Event payload
  * @param    {...*}       [args]   Any args passed are logged
  *
- * @returns  {Boolean}             Always returns `true`
+ * @returns  {true}                Always returns `true`
+ * @curried
  * @memberof core/handle
  * @public
  */
@@ -605,7 +629,7 @@ const log = handle.log = curry((message, ev, ...args) => {
  * @method   call
  * @param    {String}     method  Name of method
  *
- * @returns  {Boolean}            Returns the value returned by `method`, or `false` if the method
+ * @returns  {HandlerFunction}    Returns the value returned by `method`, or `false` if the method
  *                                does not exist
  * @memberof core/handle
  * @public
@@ -646,7 +670,8 @@ const call = function (method) {
  * @param    {Function}  handler  Handler to call with the new event payload
  * @param    {...*}      [args]   Additional args passed to both `adapter` and `handler`
  *
- * @returns  {Object}             New event payload
+ * @returns  {HandlerFunction}    New event payload
+ * @curried
  * @memberof core/handle
  * @public
  */