+++ packages/ui/VirtualList/VirtualListBase.js
@@ -1061,11 +1061,15 @@ const VirtualListBaseFactory = (type) => {
 		getScrollWidth = () => (this.isPrimaryDirectionVertical ? this.scrollBounds.clientWidth : this.getVirtualScrollDimension())
 
 		getVirtualScrollDimension = () => {
-			const
-				{dimensionToExtent, primary, curDataSize} = this,
-				{spacing} = this.props;
+			if (this.props.itemSizes) {
+				return this.props.itemSizes.reduce((total, size, index) => (total + size + (index > 0 ? this.props.spacing : 0)), 0);
+			} else {
+				const
+					{dimensionToExtent, primary, curDataSize} = this,
+					{spacing} = this.props;
 
-			return (Math.ceil(curDataSize / dimensionToExtent) * primary.gridSize) - spacing;
+				return (Math.ceil(curDataSize / dimensionToExtent) * primary.gridSize) - spacing;
+			}
 		}
 
 		syncClientSize = () => {
@@ -1326,14 +1330,14 @@ VirtualListBaseNative.displayName = 'ui:VirtualListBaseNative';
  * @public
  */
 
-const ScrollableVirtualList = (props) => (
+const ScrollableVirtualList = ({role, ...rest}) => (
 	<Scrollable
-		{...props}
-		childRenderer={({initChildRef, ...rest}) => ( // eslint-disable-line react/jsx-no-bind
+		{...rest}
+		childRenderer={({initChildRef, ...childRest}) => ( // eslint-disable-line react/jsx-no-bind
 			<VirtualListBase
-				{...rest}
+				{...childRest}
 				itemsRenderer={({cc, itemContainerRef}) => ( // eslint-disable-line react/jsx-no-bind
-					cc.length ? <div ref={itemContainerRef} role="list">{cc}</div> : null
+					cc.length ? <div ref={itemContainerRef} role={role}>{cc}</div> : null
 				)}
 				ref={initChildRef}
 			/>
@@ -1342,21 +1346,23 @@ const ScrollableVirtualList = (props) => (
 );
 
 ScrollableVirtualList.propTypes = {
-	direction: PropTypes.oneOf(['horizontal', 'vertical'])
+	direction: PropTypes.oneOf(['horizontal', 'vertical']),
+	role: PropTypes.string
 };
 
 ScrollableVirtualList.defaultProps = {
-	direction: 'vertical'
+	direction: 'vertical',
+	role: 'list'
 };
 
-const ScrollableVirtualListNative = (props) => (
+const ScrollableVirtualListNative = ({role, ...rest}) => (
 	<ScrollableNative
-		{...props}
-		childRenderer={({initChildRef, ...rest}) => ( // eslint-disable-line react/jsx-no-bind
+		{...rest}
+		childRenderer={({initChildRef, ...childRest}) => ( // eslint-disable-line react/jsx-no-bind
 			<VirtualListBaseNative
-				{...rest}
+				{...childRest}
 				itemsRenderer={({cc, itemContainerRef}) => ( // eslint-disable-line react/jsx-no-bind
-					cc.length ? <div ref={itemContainerRef} role="list">{cc}</div> : null
+					cc.length ? <div ref={itemContainerRef} role={role}>{cc}</div> : null
 				)}
 				ref={initChildRef}
 			/>
@@ -1365,11 +1371,13 @@ const ScrollableVirtualListNative = (props) => (
 );
 
 ScrollableVirtualListNative.propTypes = /** @lends ui/VirtualList.VirtualListBaseNative.prototype */ {
-	direction: PropTypes.oneOf(['horizontal', 'vertical'])
+	direction: PropTypes.oneOf(['horizontal', 'vertical']),
+	role: PropTypes.string
 };
 
 ScrollableVirtualListNative.defaultProps = {
-	direction: 'vertical'
+	direction: 'vertical',
+	role: 'list'
 };
 
 export default VirtualListBase;