+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [3.1.0] - 2019-09-16
+
+### Added
+
+- `spotlight` support for passing a spottable node or a container node or selector to `Spotlight.focus()`
+
 ## [3.0.1] - 2019-09-09
 
 No significant changes.